decided to go with 2 threads

+++ src/render/render_upscale.c
@@ -31,13 +31,13 @@ static void	img_size(t_rt *rt)//LEFTOFF DOESNT WORK?
 	uint16_t	i;
 
 	i = 0;
-	rt->threads[i].rdr_height = (float)(rt->threads[i].img_height) / rt->win->res_ratio;
+	rt->threads[i].rdr_height = rt->threads[i].img_height / (uint16_t)(rt->win->res_ratio);
 	printf("[%d] start_y: %hu, rdr_height: %hu\n", i, rt->threads[i].start_y, rt->threads[i].rdr_height);
 	++i;
 	while (i < THREADS - 1)
 	{
-		rt->threads[i].start_y = (uint16_t)((float)(rt->threads[i - 1].img_height) / rt->win->res_ratio) + rt->threads[i - 1].start_y;
-		rt->threads[i].rdr_height = (uint16_t)((float)(rt->threads[i].img_height) / rt->win->res_ratio) + rt->threads[i].start_y;
+		rt->threads[i].start_y = rt->threads[i - 1].img_height / (uint16_t)(rt->win->res_ratio) + rt->threads[i - 1].start_y;
+		rt->threads[i].rdr_height = rt->threads[i].img_height / (uint16_t)(rt->win->res_ratio) + rt->threads[i].start_y;
 		printf("[%d] start_y: %hu, rdr_height: %hu\n", i, rt->threads[i].start_y, rt->threads[i].rdr_height);
 		++i;
 	}
